ext4: fix kernel oops caused by spurious casefold flag
If an directory has the a casefold flag set without the casefold
feature set, s_encoding will not be initialized, and this will cause
the kernel to dereference a NULL pointer.  In addition to adding
checks to avoid these kernel oops, attempts to load inodes with the
casefold flag when the casefold feature is not enable will cause the
file system to be declared corrupted.

[Jaegeuk Kim: use EXT4_ERROR_INODE]
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit ca7f5aa776453d4f6d3eed801bb5ed978ccf3558)